Remove fetchPaginated and migrate viewer.js to viewer-index.js in webpack
